requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-Login==0.6.2
Werkzeug==2.3.7
Pillow==10.1.0
python-socketio==5.9.0
eventlet==0.33.3
cryptography==41.0.7